requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
black==22.8.0
click==8.1.3
Django==4.1.1
djangorestframework==3.13.1
flake8==5.0.4
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.1
pkg_resources==0.0.0
platformdirs==2.5.2
pycodestyle==2.9.1
pyflakes==2.5.0
pytz==2022.2.1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.3.0
